requirements.txt
flask-mongoengine
flask-admin
gunicorn
pillow
web-browser